@@ -29,12 +29,11 @@ from auto_round.export.export_to_llmcompressor.config import check_compressed_te
 from auto_round.export.utils import save_model
 from auto_round.utils import (
     SUPPORTED_LAYER_TYPES,
-    _get_packing_device,
     check_start_with_block_name,
     check_to_quantized,
     copy_python_files_from_model_cache,
-    filter_quantization_config,
     get_module,
+    get_packing_device,
     logger,
     set_module,
 )
@@ -57,7 +56,7 @@ def pack_layer(layer_name: str, model: torch.nn.Module, data_type: str, device: 
     Returns:
         None: The function modifies the model in place.
     """
-    packing_device = _get_packing_device(device)
+    packing_device = get_packing_device(device)
     layer = get_module(model, layer_name)
     if hasattr(layer, "orig_layer"):
         layer = layer.orig_layer
