@@ -22,9 +22,10 @@ class LLMDataLoader:
 class TestAutoRoundAct(unittest.TestCase):
     @classmethod
     def setUpClass(self):
-        model_name = "/tf_dataset/auto_round/models/facebook/opt-125m"
-        self.model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
-        self.tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
+        self.model_name = "/tf_dataset/auto_round/models/facebook/opt-125m"
+        self.save_dir = "./saved"
+        self.model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
+        self.tokenizer = AutoTokenizer.from_pretrained(self.model_name, trust_remote_code=True)
         self.llm_dataloader = LLMDataLoader()
 
     @classmethod
@@ -138,6 +139,129 @@ class TestAutoRoundAct(unittest.TestCase):
             autoround.model.model.decoder.layers[2].self_attn.k_proj.orig_layer.act_max.shape[0],
             int(3 * 10 * 768 / 128),
         )
+
+    def test_act_config_MXFP4_saving(self):
+        scheme = "MXFP4"
+        layer_config = {"lm_head": {"act_bits": 8, "bits": 8}, "k_proj": {"act_bits": 8, "bits": 8}}
+        autoround = AutoRound(
+            self.model_name,
+            scheme=scheme,
+            iters=2,
+            seqlen=2,
+            dataset=self.llm_dataloader,
+            layer_config=layer_config,
+        )
+        quantized_model_path = self.save_dir
+        autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_round")
+        model = AutoModelForCausalLM.from_pretrained(quantized_model_path, device_map="cpu")
+        assert "lm_head" not in model.config.quantization_config.extra_config
+
+        # check inblock layer config values
+        kproj_config = model.config.quantization_config.extra_config["model.decoder.layers.1.self_attn.k_proj"]
+        assert "act_data_type" in kproj_config.keys() and kproj_config["act_data_type"] == "mx_fp_rceil"
+        assert "act_bits" in kproj_config.keys() and kproj_config["act_bits"] == 8
+        assert "act_group_size" in kproj_config.keys() and kproj_config["act_group_size"] == 32
+        assert "act_sym" in kproj_config.keys() and kproj_config["act_sym"]
+        assert "data_type" in kproj_config.keys() and kproj_config["data_type"] == "mx_fp"
+        assert "bits" in kproj_config.keys() and kproj_config["bits"] == 8
+        assert "group_size" in kproj_config.keys() and kproj_config["group_size"] == 32
+        assert "sym" in kproj_config.keys() and kproj_config["sym"]
+        shutil.rmtree(quantized_model_path, ignore_errors=True)
+
+    def test_act_config_NVFP4_saving(self):
+        scheme = "NVFP4"
+        layer_config = {"k_proj": {"act_bits": 16, "bits": 16}}
+        autoround = AutoRound(
+            self.model_name,
+            scheme=scheme,
+            iters=2,
+            seqlen=2,
+            dataset=self.llm_dataloader,
+            layer_config=layer_config,
+        )
+        quantized_model_path = self.save_dir
+        autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_round")
+        model = AutoModelForCausalLM.from_pretrained(quantized_model_path, device_map="cpu")
+        kproj_config = model.config.quantization_config.extra_config["model.decoder.layers.1.self_attn.k_proj"]
+        assert "act_data_type" in kproj_config.keys() and kproj_config["act_data_type"] == "nv_fp4_with_static_gs"
+        assert "act_bits" in kproj_config.keys() and kproj_config["act_bits"] == 16
+        assert "act_group_size" in kproj_config.keys() and kproj_config["act_group_size"] == 16
+        assert "act_sym" in kproj_config.keys() and kproj_config["act_sym"]
+        assert "data_type" in kproj_config.keys() and kproj_config["data_type"] == "nv_fp"
+        assert "bits" in kproj_config.keys() and kproj_config["bits"] == 16
+        assert "group_size" in kproj_config.keys() and kproj_config["group_size"] == 16
+        assert "sym" in kproj_config.keys() and kproj_config["sym"]
+        shutil.rmtree(quantized_model_path, ignore_errors=True)
+
+    def test_WOQ_config_INT_saving(self):
+        scheme = "W4A16"
+        layer_config = {"k_proj": {"bits": 8}}
+        autoround = AutoRound(
+            self.model_name,
+            scheme=scheme,
+            iters=2,
+            seqlen=2,
+            sym=False,
+            dataset=self.llm_dataloader,
+            layer_config=layer_config,
+        )
+        quantized_model_path = self.save_dir
+        autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_round")
+        model = AutoModelForCausalLM.from_pretrained(quantized_model_path, device_map="cpu")
+        extra_config = model.config.quantization_config.extra_config
+
+        # check inblock layer config values
+        kproj_config = extra_config["model.decoder.layers.1.self_attn.k_proj"]
+        assert "act_data_type" in kproj_config.keys() and kproj_config["act_data_type"] == "float"
+        assert "act_bits" in kproj_config.keys() and kproj_config["act_bits"] == 16
+        assert "act_group_size" in kproj_config.keys() and kproj_config["act_group_size"] == 128
+        assert "act_sym" in kproj_config.keys() and not kproj_config["act_sym"]
+        assert "data_type" in kproj_config.keys() and kproj_config["data_type"] == "int"
+        assert "bits" in kproj_config.keys() and kproj_config["bits"] == 8
+        assert "group_size" in kproj_config.keys() and kproj_config["group_size"] == 128
+        assert "sym" in kproj_config.keys() and not kproj_config["sym"]
+        assert "act_dynamic" in kproj_config.keys() and kproj_config["act_dynamic"]
+        shutil.rmtree(quantized_model_path, ignore_errors=True)
+
+    def test_act_config_FP8_saving(self):
+        scheme = "FP8_STATIC"
+        layer_config = {
+            "lm_head": {"act_bits": 8, "bits": 8},
+            # check fp8 woq config
+            "k_proj": {
+                "bits": 8,
+                "group_size": 0,
+                "data_type": "fp",
+                "act_bits": 16,
+                "act_data_type": "fp",
+            },
+        }
+        autoround = AutoRound(
+            self.model_name,
+            scheme=scheme,
+            iters=2,
+            seqlen=2,
+            dataset=self.llm_dataloader,
+            layer_config=layer_config,
+        )
+        quantized_model_path = self.save_dir
+        autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_round")
+        from transformers import AutoConfig
+
+        extra_config = AutoConfig.from_pretrained(quantized_model_path).quantization_config["extra_config"]
+        assert "lm_head" not in extra_config
+
+        # check inblock layer config values
+        kproj_config = extra_config["model.decoder.layers.0.self_attn.k_proj"]
+        assert "act_data_type" in kproj_config.keys() and kproj_config["act_data_type"] == "fp"
+        assert "act_bits" in kproj_config.keys() and kproj_config["act_bits"] == 16
+        assert "act_group_size" in kproj_config.keys() and kproj_config["act_group_size"] == 0
+        assert "act_sym" in kproj_config.keys() and kproj_config["act_sym"]
+        assert "data_type" in kproj_config.keys() and kproj_config["data_type"] == "fp"
+        assert "bits" in kproj_config.keys() and kproj_config["bits"] == 8
+        assert "group_size" in kproj_config.keys() and kproj_config["group_size"] == 0
+        assert "sym" in kproj_config.keys() and kproj_config["sym"]
+        shutil.rmtree(quantized_model_path, ignore_errors=True)
 
 
 if __name__ == "__main__":
