@@ -12,6 +12,7 @@ from auto_round import AutoRound
 
 
 class LLMDataLoader:
+
     def __init__(self):
         self.batch_size = 1
 
@@ -21,11 +22,10 @@ class LLMDataLoader:
 
 
 class TestGGUF(unittest.TestCase):
+
     @classmethod
     def setUpClass(self):
         self.model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-0.5B-Instruct"
-        self.model_name = "Qwen/Qwen2.5-0.5B-Instruct"
-        self.model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
         self.tokenizer = AutoTokenizer.from_pretrained(self.model_name, trust_remote_code=True)
         self.llm_dataloader = LLMDataLoader()
 
@@ -55,8 +55,7 @@ class TestGGUF(unittest.TestCase):
     def test_q4_0(self):
         bits, group_size, sym = 4, 32, True
         autoround = AutoRound(
-            self.model,
-            self.tokenizer,
+            self.model_name,
             bits=bits,
             group_size=group_size,
             sym=sym,
@@ -103,8 +102,7 @@ class TestGGUF(unittest.TestCase):
     def test_func(self):
         bits, group_size, sym = 4, 128, True
         autoround = AutoRound(
-            self.model,
-            self.tokenizer,
+            self.model_name,
             # bits=bits,
             # group_size=group_size,
             # sym=sym,
@@ -331,10 +329,81 @@ class TestGGUF(unittest.TestCase):
         for file_name in os.listdir(quantized_model_path):
             file_size = os.path.getsize(os.path.join(quantized_model_path, file_name)) / 1024**2
             if file_name == "mmproj-model.gguf":
-                self.assertAlmostEqual(file_size, 2535, delta=1.0)
+                self.assertAlmostEqual(file_size, 2537, delta=5.0)
             else:
-                self.assertAlmostEqual(file_size, 892, delta=1.0)
+                self.assertAlmostEqual(file_size, 892, delta=5.0)
         shutil.rmtree("./saved", ignore_errors=True)
+
+    def test_qtype_setting(self):
+        # Qwen2.5-0.5B-Instruct no output, token_embed q6_k fallbakc to q8_0 336M
+        # Qwen3-0.6B output q6_k, token_embed q4_0  448M
+        # Qwen3-8B output q6_k, token_embed q4_0 4.5G
+        # Llama-3.2-1B-Instruct o output, token_embed q6_k 736M
+        from auto_round.compressors.utils import set_layer_config
+        from auto_round.export.export_to_gguf.config import ModelType
+
+        model_name = "/tf_dataset/auto_round/models/Qwen/Qwen2.5-0.5B-Instruct"
+        ar = AutoRound(model=model_name, scheme="gguf:q4_0", iters=0)
+        ar.formats = ["gguf:q4_0"]
+        ar.layer_config, _, _ = set_layer_config(
+            ar.model,
+            ar.layer_config,
+            ar.scheme,
+            ar.scale_dtype,
+            ar.supported_types,
+            ar.inner_supported_types,
+            ar.quant_block_list,
+            ar.fp_layers,
+            ar.quant_lm_head,
+            enable_gguf_official_mixed=True,
+            is_mllm=ar.mllm,
+        )
+        self.assertTrue(ar.layer_config["model.embed_tokens"]["bits"] == 8)
+        self.assertTrue("lm_head" not in ar.layer_config)
+
+        model_name = "Qwen/Qwen3-0.6B"
+        ar = AutoRound(model=model_name, scheme="gguf:q4_0", iters=0)
+        ar.formats = ["gguf:q4_0"]
+        ar.layer_config, _, _ = set_layer_config(
+            ar.model,
+            ar.layer_config,
+            ar.scheme,
+            ar.scale_dtype,
+            ar.supported_types,
+            ar.inner_supported_types,
+            ar.quant_block_list,
+            ar.fp_layers,
+            ar.quant_lm_head,
+            enable_gguf_official_mixed=True,
+            is_mllm=ar.mllm,
+        )
+        self.assertTrue(ar.layer_config["model.embed_tokens"]["bits"] == 4)
+        self.assertTrue(ar.layer_config["lm_head"]["bits"] == 6 and ar.layer_config["lm_head"]["super_bits"] == 8)
+
+        layer_config = {
+            "model.embed_tokens": {"bits": 6, "super_bits": 8},
+            "lm_head": {"bits": 4},
+        }
+        ar = AutoRound(model=model_name, scheme="gguf:q4_0", iters=0, layer_config=layer_config)
+        ar.formats = ["gguf:q4_0"]
+        ar.layer_config, _, _ = set_layer_config(
+            ar.model,
+            ar.layer_config,
+            ar.scheme,
+            ar.scale_dtype,
+            ar.supported_types,
+            ar.inner_supported_types,
+            ar.quant_block_list,
+            ar.fp_layers,
+            ar.quant_lm_head,
+            enable_gguf_official_mixed=True,
+            is_mllm=ar.mllm,
+        )
+        self.assertTrue(ar.layer_config["lm_head"]["bits"] == 4)
+        self.assertTrue(
+            ar.layer_config["model.embed_tokens"]["bits"] == 6
+            and ar.layer_config["model.embed_tokens"]["super_bits"] == 8
+        )
 
 
 if __name__ == "__main__":
