@@ -10,6 +10,7 @@ from auto_round import schemes as ar_schemes
 from auto_round.experimental import qmodules as ar_qmodules
 from auto_round.export.export_to_autoround import AutoRoundFormat
 from auto_round.export.export_to_autoround import qlinear_fp as ar_qlinear_fp
+from auto_round.testing_utils import has_module
 
 testing_schemes = [AutoRoundFormat.MXFP8.value, AutoRoundFormat.MXFP4.value, AutoRoundFormat.NVFP4.value]
 QMODULE_MAPPING = {
@@ -17,14 +18,6 @@ QMODULE_MAPPING = {
     AutoRoundFormat.MXFP4.value: ar_qmodules.MXFP4QuantLinear,
     AutoRoundFormat.NVFP4.value: ar_qmodules.NVFP4QuantLinear,
 }
-
-
-def has_module(model: torch.nn.Module, target_module_type: torch.nn.Module) -> bool:
-    """Check if the model contains a specific module type."""
-    for _, module in model.named_modules():
-        if isinstance(module, target_module_type):
-            return True
-    return False
 
 
 @pytest.mark.parametrize("scheme", testing_schemes)
